+++ ignite/services/chain/simulate.go
@@ -7,10 +7,10 @@ import (
 )
 
 type simappOptions struct {
-	enabled     bool
-	config      simulation.Config
-	period      uint
-	genesisTime int64
+	simulationTestName string
+	enabled            bool
+	config             simulation.Config
+	genesisTime        int64
 }
 
 func newSimappOptions() simappOptions {
@@ -19,20 +19,12 @@ func newSimappOptions() simappOptions {
 			Commit: true,
 		},
 		enabled:     true,
-		period:      0,
 		genesisTime: 0,
 	}
 }
 
 // SimappOption provides options for the simapp command.
 type SimappOption func(*simappOptions)
-
-// SimappWithPeriod allows running slow invariants only once every period assertions.
-func SimappWithPeriod(period uint) SimappOption {
-	return func(c *simappOptions) {
-		c.period = period
-	}
-}
 
 // SimappWithGenesisTime allows overriding genesis UNIX time instead of using a random UNIX time.
 func SimappWithGenesisTime(genesisTime int64) SimappOption {
@@ -45,6 +37,13 @@ func SimappWithGenesisTime(genesisTime int64) SimappOption {
 func SimappWithConfig(config simulation.Config) SimappOption {
 	return func(c *simappOptions) {
 		c.config = config
+	}
+}
+
+// SimappWithSimulationTestName allows to set the simulation test name.
+func SimappWithSimulationTestName(name string) SimappOption {
+	return func(c *simappOptions) {
+		c.simulationTestName = name
 	}
 }
 
@@ -62,9 +61,9 @@ func (c *Chain) Simulate(ctx context.Context, options ...SimappOption) error {
 	}
 	return commands.Simulation(ctx,
 		c.app.Path,
+		simappOptions.simulationTestName,
 		simappOptions.enabled,
 		simappOptions.config,
-		simappOptions.period,
 		simappOptions.genesisTime,
 	)
 }